move add_bn_layer to Core(ABC)
