+++ neuralpredictors/layers/cores/base.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from collections import OrderedDict
 
 from torch import nn
 
@@ -7,6 +8,10 @@ class Core(ABC):
     """
     Base class for the core models, taking 2d inputs and computing nonlinear features.
     """
+
+    def __init__(self) -> None:
+        super().__init__()
+        self.set_batchnorm_type()
 
     def initialize(self):
         """
@@ -28,6 +33,36 @@ class Core(ABC):
             nn.init.xavier_normal_(m.weight.data)
             if m.bias is not None:
                 m.bias.data.fill_(0)
+
+    @abstractmethod
+    def set_batchnorm_type(self):
+        """
+        Set batchnorm_layer_cls, bias_layer_cls, scale_layer_cls class attributes
+        """
+        self.batchnorm_layer_cls = None
+        self.bias_layer_cls = None
+        self.scale_layer_cls = None
+
+    def add_bn_layer(self, layer: OrderedDict, layer_idx: int):
+        for attr in ["batch_norm", "hidden_channels", "independent_bn_bias", "momentum"]:
+            if not hasattr(self, attr):
+                raise NotImplementedError(f"Subclasses must have a `{attr}` attribute.")
+
+        if self.batch_norm[layer_idx]:
+            hidden_channels = self.hidden_channels[layer_idx]
+
+            if self.independent_bn_bias:
+                layer["norm"] = self.batchnorm_layer_cls(hidden_channels, momentum=self.momentum)
+                return
+
+            bias = self.bias[layer_idx]
+            scale = self.batch_norm_scale[layer_idx]
+
+            layer["norm"] = self.batchnorm_layer_cls(hidden_channels, momentum=self.momentum, affine=bias and scale)
+            if bias and not scale:
+                layer["bias"] = self.bias_layer_cls(hidden_channels)
+            elif not bias and scale:
+                layer["scale"] = self.scale_layer_cls(hidden_channels)
 
     @abstractmethod
     def regularizer(self):
